owner exceptions

+++ LearnSpringBoot/CRUD-taskPad/taskpad/src/main/java/com/taskpad/taskpad/app/services/OwnerService.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.taskpad.taskpad.app.dto.OwnerDTO;
+import com.taskpad.taskpad.app.exceptions.NotFoundException;
 import com.taskpad.taskpad.app.mapper.OwnerMapper;
 import com.taskpad.taskpad.app.models.Owner;
 import com.taskpad.taskpad.app.repository.OwnerRep;
@@ -18,7 +19,7 @@ public class OwnerService {
 
     @Transactional(readOnly = true)
     public Owner getOwnerById(Integer id){   
-        return ownerDB.getReferenceById(id);
+        return ownerDB.findById(id).orElseThrow(NotFoundException::new);
     }
     
     @Transactional(readOnly = true)
@@ -27,12 +28,14 @@ public class OwnerService {
     }
     
     @Transactional
-    public void addOwner(OwnerDTO newOwner){
-        ownerDB.save(OwnerMapper.OwnerDTOtoEntity(newOwner));
+    public Owner addOwner(OwnerDTO newOwner){
+        return ownerDB.save(OwnerMapper.OwnerDTOtoEntity(newOwner));
     }
     
     @Transactional
     public void deleteOwner(Integer id){
+        if(!ownerDB.existsById(id)) throw new NotFoundException("Owner not Found");
+
         ownerDB.deleteById(id);
     }
